requirements.txt
Flask==2.3.2
openai==0.27.0
pdfminer.six==20201018
python-dotenv==1.0.0
werkzeug==2.3.7